==> YF_RealPerimeters/Step01_OrganizeShpData_MathesYFFires.Rmd <==
---
title: "Organize shapefile data"
author: "Kayla Mathes" 
date: "204-11-22"
output: html_document
---

# Overview

# Packages
```{r}
library(terra)
library(gdalUtilities)
library(sf)
library(tidyverse)
```

# Data

## Using the 9 Yukon Flats Fires


```{r}
options(scipen = 999)

##Read in the YF fire dataframe to select specific fires for the control trial 

YFfires = st_read("/Users/kmathes/Desktop/FireDataSelection/BuiltFireDatasets/YukonFlatsFires.csv")


##Read in Fire point data 
AlaskaFirePoints=st_read("/Users/kmathes/Desktop/DATA/AlaskaFireHistory_Points_NAD83_geojson.json")

AlaskaFirePerimeters = st_read("/Users/kmathes/Desktop/DATA/AlaskaFireHistory_Polygons_AKAlbersNAD83_geojson.json")

##Create a day, month and year column 
AlaskaFirePoints <- AlaskaFirePoints%>%
  mutate(DISCOVERYDATETIME = DISCOVERYDATETIME/1000)%>%
  mutate(CONTROLDATETIME = CONTROLDATETIME/1000)%>%
  mutate(OUTDATE = OUTDATE/1000)%>%
  mutate(LASTUPDATETIME = LASTUPDATETIME/1000)


###Transform all the Date/time columns into POSIXct class 
AlaskaFirePoints$DISCOVERYDATETIME <- as.POSIXct(AlaskaFirePoints$DISCOVERYDATETIME, origin = "1970-01-01")
AlaskaFirePoints$CONTROLDATETIME <- as.POSIXct(AlaskaFirePoints$CONTROLDATETIME, origin = "1970-01-01")
AlaskaFirePoints$OUTDATE <- as.POSIXct(AlaskaFirePoints$OUTDATE, origin = "1970-01-01")
AlaskaFirePoints$LASTUPDATETIME <- as.POSIXct(AlaskaFirePoints$LASTUPDATETIME, origin = "1970-01-01")


AlaskaFirePoints <- AlaskaFirePoints%>%
    dplyr::mutate(fireYr = lubridate::year(DISCOVERYDATETIME), 
                month = lubridate::month(DISCOVERYDATETIME), 
                day = lubridate::day(DISCOVERYDATETIME), 
                date = as.character(as.Date(DISCOVERYDATETIME, format = "%Y-%m-%D")))

AlaskaFirePoints_sub <- AlaskaFirePoints%>%
  filter((NAME == "Medicine Lake" & RECORDNUMBER == "211")|  
        (NAME == "Chloya Lakes" & RECORDNUMBER == "302")|
        (NAME == "Marten Creek" & RECORDNUMBER == "216")|
        (NAME == "Chahalie" & RECORDNUMBER == "340")|
        (NAME == "Discovery Creek" & RECORDNUMBER == "451")|
         (NAME == "Vunle Lakes" & RECORDNUMBER == "322")| 
            (NAME == "Burman Lake" & RECORDNUMBER == "373")|
          (NAME == "Kuranakh" & RECORDNUMBER == "305")|
          (NAME == "Little Mosquito " & RECORDNUMBER == "593"))%>%
  dplyr::select(RECORDNUMBER,NAME,ID, fireYr, date, month, day,  MGMTORGID, MGMTOPTIONID, LATITUDE, LONGITUDE, ESTIMATEDTOTALACRES, geometry)%>%
  mutate(burnarea_m2 = ESTIMATEDTOTALACRES*4046.86)%>%
  mutate(buffer_m = sqrt((burnarea_m2/pi)))%>%
  mutate(buffer_m_adjust = case_when(NAME == "Burman Lake" ~ 15, 
                                     TRUE ~ buffer_m))

###Create a circular buffer of the burned area using the estimated total acres. This is not exact because it is not the actual burned locations, but it will allow me to run the combustion model for right now using a hypothetical burned area. 
##I had to increase the buffer area for burman lake to 15m. The tiff file was not working and I think this is because it was less than 30 m 

AlaskaFirePoints_sub_perimeters <- st_buffer(AlaskaFirePoints_sub, AlaskaFirePoints_sub$buffer_m_adjust) 


```



# Write to file

```{r eval=FALSE, include=FALSE}
st_write(AlaskaFirePoints_sub_perimeters, "../OutputYFFires/YFFiresPerimeters.shp", driver  = "ESRI Shapefile", append = FALSE)

```
