==> YF_RealPerimeters/Step13_CombustionModel_MathesYFFires.Rmd <==
---
title: "Combustion Model"
author: "Anna Talucci"
date: '2023-08-03'
output: html_document
---

# Overview

Trouble shooting
https://stackoverflow.com/questions/30097730/error-when-using-predict-on-a-randomforest-object-trained-with-carets-train

# Library
```{r}
library(tidyverse)
library(ranger)
library(e1071)
library(arrow)
```
# Models
```{r}
above_model <- readRDS("../Retraining/out_path_above_adjusted3/full_model_ranger.rds")
below_model <- readRDS("../Retraining/out_path_below_adjusted3/full_model_ranger.rds")

```




```{r}
alldata = read.csv("../YF_RealPerimeters/OutputYFFires/DataCombustionModel.csv")

alldata <- alldata%>%
  rename(burn_year = ID2)%>%
  rename(Tree.cover = tree_canopy_cover)%>%
  rename(build_up_index = build_up_index_ave)%>%
  rename(burning_index = burning_index_ave)%>%
  rename(drought_code = drought_code_ave)%>%
  rename(drought_factor = drought_factor_ave)%>%
  rename(duff_moisture_code = duff_moisture_code_ave)%>%
  rename(energy_release_component = energy_release_component_ave)%>%
  rename(fine_fuel_moisture_code = fine_fuel_moisture_code_ave)%>%
  rename(fire_daily_severity_rating = fire_daily_severity_rating_ave)%>%
 rename(fire_danger_index = fire_danger_index_ave)%>%
  rename(fire_weather_index = fire_weather_index_ave)%>%
  rename(ignition_component = ignition_component_ave)%>%
   rename(initial_fire_spread_index = initial_fire_spread_index_ave)%>%
  rename(keetch_byram_drought_index = keetch_byram_drought_index_ave)%>%
   rename(spread_component = spread_component_ave)
              
          
                       
                         
                          
                          
                                     
                                    
                                       
                                          

  

```
``

# Read in predictive data from stefano

```{r}
#abovefed <- read_parquet("../From_Anna/TrainingDataFrame/abovefed_for_predict.parquet")
#head(abovefed)
```

# Training Data

Original model training data from Stefano Potter. 

```{r}
above_train =  read.csv("../Retraining/new_aboveground_data_training_fortraining.csv", header=TRUE, sep = ",", strip.white = TRUE)
```

```{r}
below_train =  read.csv("../Retraining/new_belowground_data_training_fortraining.csv", header=TRUE, sep = ",", strip.white = TRUE)
```

```{r}
above_importance =  read.csv("../Retraining/out_path_above_adjusted3/rfe_importance.csv", header=TRUE, sep = ",", strip.white = TRUE)
```

```{r}
below_importance =  read.csv("../Retraining/out_path_below_adjusted3/rfe_importance.csv", header=TRUE, sep = ",", strip.white = TRUE)
```

# Check with AboveFed Training data

```{r}
##setdiff(names(abovefed), names(alldata))
```

```{r}
setdiff(names(below_train), names(alldata))
```

```{r}
setdiff(names(above_train), names(alldata))
```

```{r}
setdiff(names(above_train), names(below_train))
```
```{r}
setdiff(below_importance$Variables, names(alldata))
```

```{r}
setdiff(above_importance$Variables, names(alldata))
```

```{r}
below_importance
```

# Select only importance variables and add column for combustion for training

Based on Stefano Potter's original model.

```{r}
above_train2 <- above_train %>% dplyr::select(above_importance$Variables)
below_train2 <- below_train %>% dplyr::select(below_importance$Variables)
above_train2$above.carbon.combusted = above_train$above.carbon.combusted
below_train2$below.ground.carbon.combusted = below_train$below.ground.carbon.combusted
```


# Training models 

Training each model with optimized parameters from Stefano Potter's ABOVEFed Model

```{r}
model_ranger_above <- ranger(above.carbon.combusted ~., data = above_train2, mtry = above_model$bestTune$mtry, splitrule = above_model$bestTune$splitrule, min.node.size = above_model$bestTune$min.node.size)
```

```{r}
model_ranger_below <- ranger(below.ground.carbon.combusted ~., data = below_train2, mtry = below_model$bestTune$mtry, splitrule = below_model$bestTune$splitrule, min.node.size = below_model$bestTune$min.node.size)
```

```{r}
rownames(below_model$importance)
below_model$bestTune$splitrule
below_model$bestTune$mtry
below_model$bestTune$min.node.size
```

# Predict combustion with all independent variables
These predictive models use all independent variables.

## Drop NA

Must remove NAs or it will not predict. 

```{r}
( alldata = alldata %>% drop_na() )
```
## select variables of interest

Following importance variables from Stefano's original model. 

```{r}
( alldataAbove = alldata %>% dplyr::select(above_importance$Variables) )
( alldataBelow = alldata %>% dplyr::select(below_importance$Variables) )
```

## Aboveground preditions 

```{r}
pred_above = predict(model_ranger_above, alldataAbove)
```

```{r}
AbovepredResults <- data.frame(predict(model_ranger_above, alldataAbove))
AbovepredResults
```

```{r}
above_pred <- cbind(alldata, AbovepredResults)
above_pred
```

```{r eval=FALSE, include=FALSE}
write_csv(above_pred, "../YF_RealPerimeters/OutputYFFires/AbovegroundCombustionRetrained.csv")
```


## Below Combustion
 
```{r}
pred_below = predict(model_ranger_below, alldataBelow)
```

```{r}
BelowpredResults <- data.frame(predict(model_ranger_below, alldataBelow))

```

```{r}
below_pred <- cbind(alldata, BelowpredResults)
```

```{r eval=FALSE, include=FALSE}
write_csv(below_pred, "../YF_RealPerimeters/OutputYFFires/BelowgroundCombustionRetrained.csv")
```
